requirements.txt
cohere
groq
rich
python-dotenv
google-generativeai
requests
colorama
anthropic
openai>=1.13.3
googlesearch-python
mistralai
nest_asyncio
pydantic>=2.4.2
mcp>=1.0.0
duckduckgo-search
HelpingAI
